Added Ptrs

+++ Signal.cpp
@@ -3,7 +3,7 @@
 Signal::Signal(){
    prop_speed = 300000000;
 };
-double Signal::get_Signal(double distance,unsigned int time){
+double Signal::get_Signal(double distance,double time){
    //Default return 1 hertz Signal
    return cos(distance/prop_speed-time);
 };
